Rename module dir
