+++ iml-gui/crate/src/page/snapshot/take.rs
